Update README.md
Detection of Lung Infection using CNN

Artificial Intelligence has evolved a lot and is currently able to solve problems that are very complex and require human specialization. One such area is healthcare.
 
A lot of research happens every day to use deep learning for the betterment of humanity, and one such is healthcare.