Update README.md
Detection of Lung Infection using Keras

Artificial Intelligence has evolved a lot and is currently able to solve problems that are very complex and require human specialization. A lot of research happens every day to use deep learning for the betterment of humanity, and one such is healthcare. I will be building a CNN model that can classify lung
disease according to the medical imagery data provided.